requirements.txt
langgraph>=0.2.0
anthropic>=0.34.0
pyyaml>=6.0
